requirements.txt
asgiref==3.4.1
Django==3.0.1
djangorestframework==3.11.0
psycopg2-binary==2.8.6
pytz==2021.3
sqlparse==0.4.2
